signal an error on missing SBT build spec file
sbt correctly functions and generates ENSIME config files even when a
`build.sbt` or `build.scala` file is missing.  In this situation,
however, the `sbt:find-root` doesn't recognize the project as
such (returns nil), causing cryptic error messages to be signaled to the
user while trying to run an sbt command. The typical message was
"wrong-type-argument stringp nil".

After this commit, a more helpful error message is given to the user,
with a recommendation to create a `build.sbt`/`build.scala`.