Fix REPL for out-of-band output
Without this fix any out-of-band output (i.e. output from background threads) would get appended to the prompt and thus locking the end of buffer as read-only so that you could not insert commands to the line any more.

This fix changes the handing of output so that it is always inserted before the current prompt.