feat: add variable for customizing how SBT gets displayed
I prefer to have SBT show up in a small window at the bottom of my screen. This
gives me a hook to make that happen, while not boxing anyone else into that; it
is a non-breaking change.

I hope this fits in a way inconsistent with the existing design. I used a
`defvar` rather than a `defcustom`, which I saw for a lot of things, because
this new variable is unlikely to be useful without writing a bit of ELisp.

This change is necessary because `display-buffer-overriding-action` is not for
user customization, and `display-buffer-alist` gets clobbered if you have
purpose-based window display via `purpose--action-function`. That is to say, I
pursued those other routes first, and it didn't work, which is why I introduced
this change.